Refactor grammar.rb that separate Symbol class from lrama/grammar.rb
