+++ lib/lrama.rb
@@ -11,5 +11,6 @@ require "lrama/rule"
 require "lrama/state"
 require "lrama/states"
 require "lrama/states_reporter"
+require "lrama/symbol"
 require "lrama/version"
 require "lrama/warning"
